✨ feat(registry): 注册 registry 命令到 CLI
- 创建 registry/mod.rs 注册 9 个子命令（含嵌套 package/tag 子命令组）
- 注册 registry 模块到 commands/mod.rs 和 main.rs
- 修复 format_bytes 参数类型
- 编译通过

+++ src/commands/registry/list.rs
@@ -68,7 +68,7 @@ pub async fn run(ctx: &AppContext, args: &ListArgs) -> Result<()> {
             r.name.clone(),
             r.kind.clone(),
             r.pkg_count.to_string(),
-            format_bytes(r.used_size),
+            format_bytes(r.used_size as i64),
             visibility,
             last_push,
         ]);
